refactor: channel / chat route 분리
