+++ src/app/(providers)/(root)/[workspaceId]/channel/(home)/_components/ChannelItem/ChannelItem.tsx
