+++ src/app/(providers)/(root)/[workspaceId]/channel/(home)/_hooks/useChannelHandlers.ts
@@ -1,9 +1,9 @@
 import { useQueryClient } from '@tanstack/react-query';
 import { useCallback } from 'react';
-import { QUERY_KEYS } from '../../_constants/constants';
 import { updateChatChannels } from '../_utils/updateChatChannels';
 import { ChatSubscribePayloadProps } from '@/types/chat';
 import { GetChannelsResponse } from '@/types/channel';
+import { QUERY_KEYS } from '../../_constants/constants';
 
 type WorkspaceInfoProps = {
   workspaceId: number;
@@ -14,7 +14,7 @@ export const useChannelHandlers = () => {
 
   const handleChatInserts = useCallback(({ workspaceId }: WorkspaceInfoProps) => {
     return ({ new: payload }: { new: ChatSubscribePayloadProps }) => {
-      queryClient.setQueryData(QUERY_KEYS.CHANNELS({ workspaceId }), (prev: GetChannelsResponse[]) => {
+      queryClient.setQueryData(QUERY_KEYS.CHANNELS(workspaceId), (prev: GetChannelsResponse[]) => {
         return updateChatChannels(prev, payload);
       });
     };
@@ -23,7 +23,7 @@ export const useChannelHandlers = () => {
   const handleChannelUserUpdates = useCallback(({ workspaceId }: WorkspaceInfoProps) => {
     return () => {
       queryClient.invalidateQueries({
-        queryKey: QUERY_KEYS.CHANNELS({ workspaceId })
+        queryKey: QUERY_KEYS.CHANNELS(workspaceId)
       });
     };
   }, []);
