+++ src/app/(providers)/(root)/[workspaceId]/channel/(home)/_utils/updateChatChannels.ts
