+++ src/app/(providers)/(root)/[workspaceId]/channel/(home)/layout.tsx
@@ -22,7 +22,7 @@ const ChatLayout = ({ children }: StrictPropsWithChildren) => {
 
 const VideoChatButton = ({ workspaceId }: { workspaceId: number }) => {
   return (
-    <Link href={`/${workspaceId}/chat/add?type=video`}>
+    <Link href={`/${workspaceId}/channel/add?type=video`}>
       <Video48Icon className="text-grey700Black stroke-current w-6 h-6" />
     </Link>
   );
@@ -30,7 +30,7 @@ const VideoChatButton = ({ workspaceId }: { workspaceId: number }) => {
 
 const MessageChatButton = ({ workspaceId }: { workspaceId: number }) => {
   return (
-    <Link href={`/${workspaceId}/chat/add?type=chat`}>
+    <Link href={`/${workspaceId}/channel/add?type=chat`}>
       <MessagePlusIcon className="text-grey700Black stroke-current w-6 h-6" />
     </Link>
   );
