+++ src/app/(providers)/(root)/[workspaceId]/channel/_hooks/useChannelMutation.ts
@@ -1,26 +1,7 @@
 import api from '@/api';
-import type { ChannelInsertType } from '@/types/channel';
-import type { CreateChannelUsersProps } from '@/types/channelUser';
-import type { CreateChatMessageProps } from '@/types/chat';
+import { ChannelInsertType } from '@/types/channel';
+import { CreateChannelUsersProps } from '@/types/channelUser';
 import { useMutation } from '@tanstack/react-query';
-
-type UseChatMessageProps = Pick<CreateChatMessageProps, 'channel_id'> & {
-  onSuccess?: () => void;
-};
-type UseChatMessageMutationProps = Omit<CreateChatMessageProps, 'channel_id'>;
-
-export const useMutationChatMessage = ({ channel_id, ...props }: UseChatMessageProps) => {
-  return useMutation({
-    mutationFn: ({ content, type }: UseChatMessageMutationProps) => {
-      return api.chat.createChatMessage({
-        channel_id,
-        content,
-        type
-      });
-    },
-    ...props
-  });
-};
 
 export const useMutationCreateChannelUsers = () => {
   return useMutation({
