+++ src/app/(providers)/(root)/[workspaceId]/channel/add/_components/SearchResults.tsx
