+++ src/app/(providers)/(root)/[workspaceId]/chat/[id]/_components/ContextMenu/ContextMenu.tsx
@@ -7,9 +7,8 @@ import { supabase } from '@/utils/supabase/supabaseClient';
 import { CHAT_TYPE } from '@/constants/chat';
 import clsx from 'clsx';
 import { useSnackBar } from '@/providers/SnackBarContext';
-import { useMutationChatMessage } from '../../../_hooks/useMutationChat';
+import { useMutationChatMessage } from '../../../_hooks/useChatMutation';
 import { useParams } from 'next/navigation';
-import { useWorkspaceUserId } from '@/hooks/useWorkspaceUserId';
 
 const ContextMenu = () => {
   const { id } = useParams();
