+++ src/app/(providers)/(root)/[workspaceId]/chat/[id]/_components/MessageSender/MessageSender.tsx
@@ -1,5 +1,5 @@
 import { useRef } from 'react';
-import { useMutationChatMessage } from '../../../_hooks/useMutationChat';
+import { useMutationChatMessage } from '../../../_hooks/useChatMutation';
 import MessageTextarea from '../MessageTextarea';
 import UtilsMenu from '../UtilsMenu';
 import ContextMenu from '../ContextMenu';
