+++ src/app/(providers)/(root)/[workspaceId]/chat/[id]/_components/Messages/Messages.tsx
@@ -3,7 +3,7 @@
 import { useEffect, useRef } from 'react';
 import Chats from '../Chats';
 import { useParams } from 'next/navigation';
-import { useGetChatMessages, useGetUsersInChannel } from '../../../_hooks/useQueryChat';
+import { useGetChatMessages, useGetUsersInChannel } from '../../../_hooks/useChatQuery';
 import { handleSubscribeToChat } from '../../../_utils/subscribe';
 import { MessagesWrapper } from '../MessagesContainer';
 import { useChatHandlers } from '../../_hook/useChatHandlers';
