+++ src/app/(providers)/(root)/[workspaceId]/chat/[id]/_components/UtilsMenu/UtilsMenu.tsx
@@ -3,7 +3,7 @@ import PaperClipIcon from '@/icons/paperclip.svg';
 import VideoIcon from '@/icons/video.svg';
 import { mbToBytes } from '@/utils/file';
 import FileButton from '../FileButton';
-import { useMutationChatMessage } from '../../../_hooks/useMutationChat';
+import { useMutationChatMessage } from '../../../_hooks/useChatMutation';
 import { supabase } from '@/utils/supabase/supabaseClient';
 import { ChatType } from '@/types/chat';
 import { useSnackBar } from '@/providers/SnackBarContext';
