--- src/app/(providers)/(root)/[workspaceId]/chat/_utils/getChannelName.ts
@@ -1,9 +1,0 @@
-import { GetUsersInChannelResponse } from '@/types/channel';
-
-export const getDmChannelName = (usersInChannel: GetUsersInChannelResponse) => {
-  const result = Object.keys(usersInChannel).map((key) => {
-    return usersInChannel[key].name;
-  });
-
-  return result.length > 1 ? null : result[0];
-};
