+++ src/constants/realtime.ts
@@ -1,0 +1,5 @@
+export const REALTIME_CHANNEL_NAME = {
+  CHANNEL_LIST: 'channel_list',
+  CHAT: 'chat',
+  CHAT_FOR_NOTICE: 'chat_for_notice'
+};
